+++ main.go
@@ -1,44 +1,15 @@
 package main
 
 import (
-	"bufio"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"runtime/debug"
-	"strings"
 )
-
-// OpItem is the struct for the output of "op item get --format json" command
-// we are only interessted in key value pairs from fields as label and value
-// to get the username and password, nothing else
-// Reference: https://support.1password.com/command-line-reference/#item-get
-type OpItem struct {
-	Label string `json:"label,omitempty"`
-	Value string `json:"value,omitempty"`
-}
-
-type OpItemList []OpItem
 
 // versioning is not yet implemented
-var (
-	prefix      string
-	opItemFlags []string
-	version     = "main"
-)
-
-// GetField returns the value of the field with the given label
-func (i OpItemList) GetField(label string) string {
-	for _, field := range i {
-		if field.Label == label {
-			return field.Value
-		}
-	}
-	return ""
-}
+var version = "main"
 
 // getVersion returns the version of the binary
 func getVersion() string {
@@ -54,68 +25,11 @@ func PrintVersion() {
 	fmt.Fprintf(os.Stderr, "git-credential-1password %s\n", getVersion())
 }
 
-// get 1password item name
-func itemName(host string) string {
-	return fmt.Sprintf("%s%s", prefix, host)
-}
-
-// build a exec.Cmd for "op item" sub command including additional flags
-func buildOpItemCommand(subcommand string, args ...string) *exec.Cmd {
-	cmdArgs := []string{"item", subcommand}
-	cmdArgs = append(cmdArgs, opItemFlags...)
-	cmdArgs = append(cmdArgs, args...)
-	return exec.Command("op", cmdArgs...)
-}
-
-// opGetItem runs "op item get --format json" command with the given name
-func opGetItem(n string) (OpItemList, error) {
-	// --fields username,password limits the output to only username and password
-	opItemGet := buildOpItemCommand("get", "--format", "json", "--fields", "username,password", n)
-	opItemRaw, err := opItemGet.CombinedOutput()
-	if err != nil {
-		return nil, fmt.Errorf("opItemGet failed with %s\n%+s", err, opItemRaw)
-	}
-
-	// marhsal the raw output to OpItem struct
-	var opItem OpItemList
-	if err = json.Unmarshal(opItemRaw, &opItem); err != nil {
-		return nil, fmt.Errorf("json.Unmarshal() failed with %s", err)
-	}
-	return opItem, nil
-}
-
-// ReadLines reads the input from stdin and returns a map of key value pairs
-func ReadLines() (inputs map[string]string) {
-	inputs = make(map[string]string)
-	// create stdin reader
-	reader := bufio.NewReader(os.Stdin)
-
-	for {
-		// line by line read from stdin
-		line, _ := reader.ReadString('\n')
-
-		// if the line is empty, break the loop
-		if line == "" || line == "\n" {
-			break
-		}
-
-		// create a slice of strings by splitting the line
-		parts := strings.SplitN(line, "=", 2)
-
-		// see if this can a key value pair
-		if len(parts) == 2 {
-			inputs[strings.TrimSpace(parts[0])] = strings.TrimSpace(parts[1])
-		} else {
-			log.Fatalf("Invalid input: %s", line)
-		}
-	}
-	return inputs
-}
-
 func main() {
 	accountFlag := flag.String("account", "", "1Password account")
 	vaultFlag := flag.String("vault", "", "1Password vault")
-	prefixFlag := flag.String("prefix", "", "1Password item name prefix")
+	flag.StringVar(&prefix, "prefix", "", "1Password item name prefix")
+	flag.StringVar(&tags, "tags", "git-credentials", "1Password item tags")
 	versionFlag := flag.Bool("version", false, "Print version")
 
 	flag.Usage = func() {
@@ -146,7 +60,6 @@ func main() {
 	}
 
 	// set global variables based on flags
-	prefix = *prefixFlag
 	if *accountFlag != "" {
 		opItemFlags = append(opItemFlags, "--account", *accountFlag)
 	}
